fix: Campo integer da migration não armazena a quantidade necessaria de caracteres de 1 cpf, foi alterado para biginteger

+++ database/migrations/2024_08_28_084636_create_caixas_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('banco');
             $table->integer('agencia');
             $table->integer('conta');
-            $table->integer('cpf')->nullable();
+            $table->bigInteger('cpf')->nullable();
             $table->timestamps();
         });
     }
